requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning libraries
scikit-learn>=1.0.0

# Deep learning libraries
tensorflow>=2.8.0
keras>=2.8.0

# Image processing (optional, for advanced image operations)
Pillow>=8.3.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
tqdm>=4.62.0
